Make sure the test covers _getProductCollection
This changes the testGetIdentities so that when test runs, it would go through into _getProductCollection call - previously, it was just returning the cached collection from the property that was set via ReflectionProperty::setValue. Now, the collection is returned from the layer-object inside the _getProductCollection, instead. 
Also, a registry-mock call('registry') is added so that a case when this is called on the product-page would be covered.